refactor(recent activity): convert to class and decouple db logic to dao layer
* Decided to not extend the BaseService and BaseRepository because the recent-activity-service has only one method and does not use/need any of the base/parent methods.
